+++ ext_tables.sql
@@ -190,7 +190,7 @@ CREATE TABLE tx_seminars_seminars (
     object_type int(11) unsigned DEFAULT '0' NOT NULL,
     title tinytext,
     topic int(11) unsigned DEFAULT '0' NOT NULL,
-    slug varchar(262) DEFAULT '' NOT NULL,
+    slug varchar(262),
     subtitle tinytext,
     categories int(11) unsigned DEFAULT '0' NOT NULL,
     teaser text,