+++ Classes/UpgradeWizards/GenerateEventSlugsUpgradeWizard.php
@@ -12,6 +12,7 @@ use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Install\Updates\DatabaseUpdatedPrerequisite;
+use TYPO3\CMS\Install\Updates\RepeatableInterface;
 use TYPO3\CMS\Install\Updates\UpgradeWizardInterface;
 
 /**
@@ -19,7 +20,7 @@ use TYPO3\CMS\Install\Updates\UpgradeWizardInterface;
  *
  * @deprecated will be removed in seminars 7.0
  */
-class GenerateEventSlugsUpgradeWizard implements UpgradeWizardInterface, LoggerAwareInterface
+class GenerateEventSlugsUpgradeWizard implements UpgradeWizardInterface, RepeatableInterface, LoggerAwareInterface
 {
     use LoggerAwareTrait;