+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
+- Document that the slug field should to be enabled for editors (#2556)
 - Make the object type for events a drop-down (#2551)
 - Make the event slug field nullable (#2546)
 - Make the event slug field shorter (#2539)