+++ CHANGELOG.md
@@ -6,7 +6,6 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
-- Add a PSR-14 event for creating/modifying event slugs (#2608)
 
 ### Changed
 
@@ -15,6 +14,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+
+## 5.4.0
+
+### Added
+- Add a PSR-14 event for creating/modifying event slugs (#2608)
 
 ### 5.3.1